pyproject.toml
[tool.poetry]
name = "starlette-sqlalchemy"
version = "0.1.0"
description = ""
authors = ["Jonathan Kim <jkimbo@gmail.com>"]

[tool.poetry.dependencies]
python = "3.11.0"
uvicorn = {extras = ["standard"], version = "0.20.0"}
SQLAlchemy = {extras = ["mypy"], version = "1.4.23"}
alembic = "1.9.4"
strawberry-graphql = "0.158.0"
fastapi = "0.92.0"
typing_extensions = "4.5.0"
aiosqlite = "0.18.0"


[tool.poetry.dev-dependencies]
black = "^21.8b0"
flake8 = "^3.9.2"
flake8-black = "^0.2.3"
flake8-bugbear = "^21.4.3"
mypy = "^0.910"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
